pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nwo-parts-gallery"
version = "0.1.0"
description = "NWO Robotics Layer 2 — Parts Gallery & Agent File Store"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "NWO Robotics", email = "dev@nworobotics.cloud" }]
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",           # Gallery HTML templates

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",         # Async PostgreSQL driver
    "alembic>=1.13.1",         # Migrations
    "pgvector>=0.3.2",         # Vector similarity search

    # Blob storage
    "boto3>=1.34.0",           # S3 / MinIO / R2
    "aiobotocore>=2.13.0",     # Async S3

    # Embeddings (for semantic search)
    "openai>=1.35.0",          # text-embedding-3-small (default)
    "anthropic>=0.28.0",       # Fallback
    "httpx>=0.27.0",           # Ollama embeddings

    # Mesh preview (thumbnail generation)
    "trimesh>=4.4.0",
    "numpy>=1.26.0",
    "pillow>=10.3.0",          # PNG thumbnail output

    # Crypto / identity
    "cryptography>=42.0.0",    # Agent key signing

    # Data / config
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "python-ulid>=2.2.0",
    "structlog>=24.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "httpx>=0.27.0",
    "ruff>=0.4.9",
    "mypy>=1.10.0",
    "factory-boy>=3.3.0",
]

[project.scripts]
nwo-gallery = "src.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
